# docs/swagger.yaml
definitions:
  commons.DataTableColumn:
    properties:
      name:
        type: string
      orderable:
        type: boolean
      searchValue:
        type: string
      searchable:
        type: boolean
    type: object
  commons.DataTableFilter:
    properties:
      field:
        type: string
      operator:
        type: string
      value:
        type: string
    type: object
  commons.DataTableOrder:
    properties:
      direction:
        type: string
      field:
        type: string
    type: object
  commons.DataTableRequest:
    properties:
      columns:
        items:
          $ref: '#/definitions/commons.DataTableColumn'
        type: array
      filters:
        items:
          $ref: '#/definitions/commons.DataTableFilter'
        type: array
      orders:
        items:
          $ref: '#/definitions/commons.DataTableOrder'
        type: array
      page:
        type: integer
      search:
        type: string
      size:
        type: integer
    type: object
...
  models.Organization:
    properties:
      approved_at:
        type: string
      approved_by:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      entity_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      is_locked:
        type: boolean
      name:
        type: string
      owner_id:
        type: string
      submitted_at:
        type: string
      submitted_by:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.ProductCategory:
    properties:
      approved_at:
        type: string
      approved_by:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      entity_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      is_locked:
        type: boolean
      name:
        type: string
      organization_id:
        type: string
      owner_id:
        type: string
      submitted_at:
        type: string
      submitted_by:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /organization/update:
    patch:
      consumes:
      - application/json
      parameters:
      - description: organization
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.Organization'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      tags:
      - Organization
...
  /product_category/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: ProductCategory
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.ProductCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product Category
  /product_category/deleteById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product Category
  /product_category/getAll:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product Category
  /product_category/getById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product Category
  /product_category/getPagination:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commons.DataTableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product Category
  /product_category/update:
    patch:
      consumes:
      - application/json
      parameters:
      - description: productCategory
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.ProductCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product Category
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"